+++ lua/terrortown/lang/tr.lua
@@ -2076,11 +2076,11 @@ L.flaregun_help_primary = "Varlığı yak"
 
 -- 2023-12-14
 L.marker_vision_owner = "Sahip: {owner}"
-L.marker_vision_distance = "Uzaklık: {distance}m"
-L.marker_vision_distance_collapsed = "{distance}m"
+L.marker_vision_distance = "Uzaklık: {distance}"
+L.marker_vision_distance_collapsed = "{distance}"
 
 L.c4_marker_vision_time = "Patlamaya: {time}"
-L.c4_marker_vision_collapsed = "{time} / {distance}m"
+L.c4_marker_vision_collapsed = "{time} / {distance}"
 
 L.c4_marker_vision_safe_zone = "Bombadan güvenli bölge"
 L.c4_marker_vision_damage_zone = "Bomba hasar bölgesi"
@@ -2794,3 +2794,14 @@ L.label_hud_show_team_name = "Rol adının yanında takım adını göstermeyi e
 --Plays sounds to distract or deceive.
 --
 --Place the radio somewhere, and then remotely interact with it to choose sounds to play.]]
+
+-- 2025-02-21
+L.length_in_meters = "{length}m"
+--L.length_in_yards = "{length}yd"
+--L.length_in_feet = "{length}ft"
+
+--L.label_distance_unit = "Preferred unit of length for distance displays"
+--L.choice_distance_unit_0 = "Inches"
+--L.choice_distance_unit_1 = "Meters"
+--L.choice_distance_unit_2 = "Yards"
+--L.choice_distance_unit_3 = "Feet"